# rttanalysis: run benchmarks with non-root user