--- a/pkg/sql/information_schema.go
+++ b/pkg/sql/information_schema.go
@@ -2830,19 +2830,13 @@ func userCanSeeDescriptor(
 		return false, nil
 	}
 
-	// Users can see objects in the database if they have connect privilege.
-	if parentDBDesc != nil {
-		if ok, err := p.HasPrivilege(ctx, parentDBDesc, privilege.CONNECT, p.User()); err != nil {
-			return false, err
-		} else if ok {
-			return true, nil
-		}
+	// Short-circuit for virtual tables, so that we avoid fetching the synthetic
+	// privileges for virtual tables in a thundering herd while populating a table
+	// like pg_class, which has a row for every table, including virtual tables.
+	if tab, ok := desc.(catalog.TableDescriptor); ok && tab.IsVirtualTable() {
+		return true, nil
 	}
-	// Also check if the user has any privilege on the descriptor. This check is
-	// done second, so that the check above can short-circuit, allowing us to
-	// avoid fetching the synthetic privileges for virtual tables in a thundering
-	// herd while populating a table like pg_class, which has a row for every
-	// table, including virtual tables.
+
 	// TODO(richardjcai): We may possibly want to remove the ability to view
 	// the descriptor if they have any privilege on the descriptor and only
 	// allow the descriptor to be viewed if they have CONNECT on the DB. #59827.
@@ -2851,6 +2845,16 @@ func userCanSeeDescriptor(
 	} else if ok {
 		return true, nil
 	}
+
+	// Users can see objects in the database if they have connect privilege.
+	if parentDBDesc != nil {
+		if ok, err := p.HasPrivilege(ctx, parentDBDesc, privilege.CONNECT, p.User()); err != nil {
+			return false, err
+		} else if ok {
+			return true, nil
+		}
+	}
+
 	return false, nil
 }