Backport 2/2 commits from #111819.

/cc @cockroachdb/release

---

### rttanalysis: run benchmarks with non-root user

The root user skips certain privilege checks, so using it in these
benchmarks would create results that don't reflect the normal experience
that most users have. The benchmark expectations are rewritten
accordingly.

At the same time, this adds new tests for queries run by npgsql that
were found to be slow.

---

### sql: improve performance of userCanSeeDescriptor

28bcaaf (and the related backports)
changed the order that checks were made when determining if a user can
view a descriptor for the purpose of populating a virtual table. It
improved the performance with virtual table descriptors, but made things
worse for other descriptors.

The performance had degraded enough to the point that the nightly npgsql
tests began failing.

This commit goes back to the previous order of checking privileges, and
also adds another check so that virtual table descriptors can still
short circuit.

```
before
BenchmarkORMQueries/npgsql_types-10         	      14	  72771417 ns/op	        60.57 roundtrips
BenchmarkORMQueries/npgsql_fields-10        	       4	 288099771 ns/op	        84.00 roundtrips

after
BenchmarkORMQueries/npgsql_types-10         	      30	  37119303 ns/op	        13.27 roundtrips
BenchmarkORMQueries/npgsql_fields-10        	      72	  16410549 ns/op	         9.056 roundtrips
```

---

Release justification: fix a performance regression that was backported
fixes https://github.com/cockroachdb/cockroach/issues/109960
Release note: None